+++ packages/amplify-cli-core/src/index.ts
@@ -25,6 +25,8 @@ export * from './cliRemoveResourcePrompt';
 export * from './cliViewAPI';
 export * from './overrides-manager';
 export * from './hooks';
+export * from './cliViewAPI';
+export * from './customPoliciesUtils'
 export * from './cliConstants';
 export * from './category-interfaces';