+++ packages/amplify-e2e-tests/src/init-special-cases/index.ts
@@ -56,7 +56,7 @@ async function initWorkflow(cwd: string, settings: { accessKeyId: string; secret
       .wait('Enter a name for the project')
       .sendCarriageReturn()
       .wait('Initialize the project with the above configuration?')
-      .sendLine('n')
+      .sendConfirmNo()
       .wait('Enter a name for the environment')
       .sendCarriageReturn()
       .wait('Choose your default editor:')