+++ packages/amplify-go-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@1.9.6...amplify-go-function-runtime-provider@1.9.7) (2021-10-10)
+
+**Note:** Version bump only for package amplify-go-function-runtime-provider
+
+
+
+
+
+## [1.9.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@1.9.5...amplify-go-function-runtime-provider@1.9.6) (2021-10-06)
+
+**Note:** Version bump only for package amplify-go-function-runtime-provider
+
+
+
+
+
+## [1.9.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@1.9.4...amplify-go-function-runtime-provider@1.9.5) (2021-09-27)
+
+**Note:** Version bump only for package amplify-go-function-runtime-provider
+
+
+
+
+
 ## [1.9.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@1.9.3...amplify-go-function-runtime-provider@1.9.4) (2021-09-18)
 
 **Note:** Version bump only for package amplify-go-function-runtime-provider