+++ packages/graphql-predictions-transformer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.24](https://github.com/aws-amplify/amplify-cli/compare/graphql-predictions-transformer@2.5.23...graphql-predictions-transformer@2.5.24) (2021-10-10)
+
+**Note:** Version bump only for package graphql-predictions-transformer
+
+
+
+
+
+## [2.5.23](https://github.com/aws-amplify/amplify-cli/compare/graphql-predictions-transformer@2.5.22...graphql-predictions-transformer@2.5.23) (2021-10-06)
+
+**Note:** Version bump only for package graphql-predictions-transformer
+
+
+
+
+
+## [2.5.22](https://github.com/aws-amplify/amplify-cli/compare/graphql-predictions-transformer@2.5.21...graphql-predictions-transformer@2.5.22) (2021-09-27)
+
+**Note:** Version bump only for package graphql-predictions-transformer
+
+
+
+
+
 ## [2.5.21](https://github.com/aws-amplify/amplify-cli/compare/graphql-predictions-transformer@2.5.20...graphql-predictions-transformer@2.5.21) (2021-09-18)
 
 **Note:** Version bump only for package graphql-predictions-transformer